Migration for slash price
I wanted a slash price to show up above the current price. A weird tactic used by alot of the e-commerce sites. Easiest way to implement this for all 8 products was to add the column to DB and now map through it as normal so I can customize each slash price